requirements.txt
flask
langchain-groq
python-dotenv
langchain-core
